moved styling objects to qastyle.jsx
