+++ client/QuestionsAnswers/qastyle.jsx
@@ -1,0 +1,18 @@
+const qHelpfulStyle = {
+  display: 'inline',
+  fontSize: 12,
+  color: 'grey',
+  marginLeft: 10,
+};
+
+const ulStyle = {
+  listStyleType: 'none',
+  margin: 0,
+  padding: 0,
+  display: 'inline',
+};
+
+module.exports = {
+  ulStyle: ulStyle,
+  qHelpfulStyle: qHelpfulStyle,
+};
